requirements.txt
keras
tensorflow
fastapi
uvicorn
opencv-python
pytesseract
numpy
python-multipart
paddleocr
paddlepaddle
pandas
python-dotenv